authorization implemented

+++ src/services/notes/controller/note-controller.js
@@ -2,14 +2,17 @@ import noteRepositories from "../repositories/note.repositories.js";
 import InvariantError from "../../../exceptions/invariant-error.js";
 import response from "../../../utils/response.js";
 import NotFoundError from "../../../exceptions/not-found-error.js";
+import AuthorizationError from "../../../exceptions/authorization-error.js";
 
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
 
   const note = await noteRepositories.createNote({
     title,
     body,
     tags,
+    owner,
   });
 
   if (!note) {
@@ -20,12 +23,23 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await noteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await noteRepositories.getNotes(owner);
   return response(res, 200, "Catatan sukses ditampilkan", notes);
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
+
   const note = await noteRepositories.getNoteById(id);
 
   if (!note) {
@@ -38,6 +52,16 @@ export const getNoteById = async (req, res, next) => {
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
+
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
 
   const note = await noteRepositories.editNote({
     id,
@@ -55,6 +79,15 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
   const deletedNote = await noteRepositories.deleteNoteById(id);
 
   if (!deletedNote) {
